+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 ## Added
 * Motion animations along UIBezierPath and CGPath
 
+## Changed
+* `.runAnimation` method were replaced with `.animate`
+
+## Fixed
+* Issue when animation could reset after calling empty animation block
+
 # Changelog for Lotus 0.2.0
 ## Added
 * Documentation on [Github Pages](lotus-docs.github.io)